+++ DigDug/Pooka.cpp
@@ -110,7 +110,7 @@ namespace cem
 
 	void Pooka::SetSpritesInvisible()
 	{
-		for (int i = 0; i < m_pSprites.size(); i++)
+		for (size_t i = 0; i < m_pSprites.size(); i++)
 		{
 			m_pSprites[i]->SetVisibility(false);
 		}
@@ -118,7 +118,7 @@ namespace cem
 
 	void Pooka::SetTexturesInvisible()
 	{
-		for (int i = 0; i < m_pTextures.size(); i++)
+		for (size_t i = 0; i < m_pTextures.size(); i++)
 		{
 			m_pTextures[i]->SetVisibility(false);
 		}
